refactor: member 패키지 내의 서비스별로 분리
