+++ src/main/java/com/souf/soufwebsite/domain/member/controller/FavoriteController.java
@@ -2,7 +2,7 @@ package com.souf.soufwebsite.domain.member.controller;
 
 import com.souf.soufwebsite.domain.member.dto.ReqDto.FavoriteMemberReqDto;
 import com.souf.soufwebsite.domain.member.dto.ResDto.MemberResDto;
-import com.souf.soufwebsite.domain.member.service.FavoriteService;
+import com.souf.soufwebsite.domain.member.service.favorite.FavoriteService;
 import com.souf.soufwebsite.global.success.SuccessResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
