+++ src/main/java/com/souf/soufwebsite/domain/member/service/favorite/FavoriteServiceImpl.java
@@ -1,4 +1,4 @@
-package com.souf.soufwebsite.domain.member.service;
+package com.souf.soufwebsite.domain.member.service.favorite;
 
 import com.souf.soufwebsite.domain.file.entity.PostType;
 import com.souf.soufwebsite.domain.file.service.FileService;
