+++ src/main/java/com/souf/soufwebsite/domain/member/service/general/MemberService.java
@@ -1,4 +1,4 @@
-package com.souf.soufwebsite.domain.member.service;
+package com.souf.soufwebsite.domain.member.service.general;
 
 import com.souf.soufwebsite.domain.file.dto.MediaReqDto;
 import com.souf.soufwebsite.domain.member.dto.ReqDto.*;
