+++ src/test/java/com/souf/soufwebsite/domain/member/service/MemberCategoryServiceTest.java
@@ -2,6 +2,7 @@ package com.souf.soufwebsite.domain.member.service;
 
 import com.souf.soufwebsite.domain.member.repository.MemberCategoryMappingRepository;
 import com.souf.soufwebsite.domain.member.repository.MemberRepository;
+import com.souf.soufwebsite.domain.member.service.general.MemberCategoryServiceImpl;
 import com.souf.soufwebsite.global.common.category.service.CategoryService;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
